requirements.txt
streamlit==1.53.1
python-dotenv==1.2.1
langchain==0.2.0
langchain-openai==0.1.3
openai==1.109.1